[refactor] UserService 구현체 제거 후 class로 변경
