+++ src/main/java/com/example/twogether/user/service/UserService.java
@@ -1,36 +1,45 @@
 package com.example.twogether.user.service;
 
+import com.example.twogether.common.error.CustomErrorCode;
+import com.example.twogether.common.exception.CustomException;
 import com.example.twogether.user.dto.SignupRequestDto;
 import com.example.twogether.user.entity.User;
+import com.example.twogether.user.entity.UserRoleEnum;
+import com.example.twogether.user.repository.UserRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.security.crypto.password.PasswordEncoder;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
-public interface UserService {
-    /**
-     * 회원 가입
-     * @param requestDto 사용자 회원 가입을 위한 요청 정보
-     * @return user 생성된 사용자 정보 반환
-     */
-    User signup(SignupRequestDto requestDto) throws IllegalAccessException;
+@Service
+@RequiredArgsConstructor
+public class UserService {
 
-    /**
-     * 유저 정보 수정
-     * @param requestDto 사용자 정보 수정을 위한 요청 정보
-     * @param user 수정할 정보를 소유하고 있는 사용자 본인에 대한 정보
-     * @return user 수정된 사용자 정보 반환
-     */
-//    User editUserInfo(EditUserRequestDto requestDto, User user);
+    private final UserRepository userRepository;
+    private final PasswordEncoder passwordEncoder;
 
-    /**
-     * 유저 정보 삭제
-     * @param userId 삭제할 사용자의 식별 ID
-     * @param user 삭제할 정보를 소유하고 있는 사용자 본인 혹은 관리자에 대한 정보
-     */
-//    void deleteUserInfo(Long userId, User user);
+    @Value("${admin.token}")
+    private String adminToken;
 
-    /**
-     * 유저 비밀번호 수정
-     * @param requestDto 사용자 비밀번호 수정을 위한 요청 정보
-     * @param user 수정할 정보를 소유하고 있는 사용자 본인에 대한 정보
-     * @return user 비밀번호 수정한 사용자 정보 반환
-     */
-//    User editUserPassword(EditPasswordRequestDto requestDto, User user);
+    @Transactional
+    public User signup(SignupRequestDto requestDto) {
+        String email = requestDto.getEmail();
+        String password = passwordEncoder.encode(requestDto.getPassword());
+
+        // 사용자 존재 여부 확인
+        findExistingUserByEmail(email);
+
+        UserRoleEnum role = UserRoleEnum.USER;
+        if (requestDto.isAdmin() && requestDto.getAdminToken().equals(adminToken)) {
+            role = UserRoleEnum.ADMIN;
+        }
+
+        return userRepository.save(requestDto.toEntity(password, role));
+    }
+
+    private void findExistingUserByEmail(String email) {
+        if(userRepository.findByEmail(email).orElse(null) != null)
+            throw new CustomException(CustomErrorCode.USER_ALREADY_EXISTS);
+    }
 }
